+++ FINLAY_SMITH/GameAssets/Vector.java
@@ -1,4 +1,4 @@
-package GameAssets;
+package FINLAY_SMITH.GameAssets;
 
 public class Vector {
     public double x;
